Adiciona a exposição de projetos no body, reorganiza as fotos em pastas separadas
