+++ frontend/src/imgs/icons/icons8-algorithm.png
