+++ frontend/src/styles/Body.css
@@ -54,7 +54,7 @@
     background-color: white;
     border-radius: 20px;
     list-style: none;
-    height: 400px;
+    height: 450px;
     width: 15%;
     padding: 10px;
     img {
