+++ frontend/src/styles/index.css
@@ -4,7 +4,7 @@ html, body, #root {
 
 body {
   margin: 0;
-  height: 300%;
+  height: 500%;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
